Various updates.
Changed units from MHz to Hz everywhere that I could find it.
Removed RF and IF functions from StationControllerKATCPClient.
Added functionality to KATCP server to read from RF Controller GUI and record RF/IF metadata.
Passed data to file writer.

Currently works only if changes are received while writing. This is a small item of concern which I'll need to fix.
I'm committing just so that the changes that I've made can be saved. It works as far as I have tested.